+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.3.0] - 2022-10-07
+## [4.3.0] - 2022-10-14
 
 ### Added
 
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   3. Redelegating tokens
   4. Claim tokens
   5. Voting on proposals
+- Add local wallet update when changed by Keplr
 - Add proposal stats
 
 ### Changed